refactor common code, use clickhouse for polling
